Refactor some files location; Move NavigationItem as individual component for the single responsibility
