+++ src/components/Dashboard/index.ts
@@ -1,7 +1,7 @@
-export { default as Sidebar } from './Sidebar';
+export { default as Sidebar } from '../Sidebar/Sidebar';
 export { default as Header } from './Header';
 export { default as PageTitle } from './PageTitle';
-export { default as HomeButton } from './HomeButton';
+export { default as HomeButton } from '../Sidebar/HomeButton';
 export { default as UserMenu } from './UserMenu';
 export { default as UserMenuDropdown } from './UserMenuDropdown';
 export { default as UserMenuTrigger } from './UserMenuTrigger';
